assignr/requirements.txt
load-dotenv==0.1.0
pandas==2.1.1
requests==2.31.0